Classifiers/requirements.txt
torch==2.3.1+cpu
torchvision==0.18.1+cpu
-f https://download.pytorch.org/whl/torch_stable.html
Pillow==11.0.0
cassandra-driver==3.29.0
Flask==2.3.3
